Module schema_engine::json_rpc_stdio

This module implements JSON-RPC over standard IO. It uses tokio for async IO, and jsonrpc_core for the JSON-RPC part.

Structs

Client 🔒

A handle to a connected client you can use to send requests.

ClientAdapter 🔒

The other side of the channels. Only used as a handle to be passed into run_with_client().

Enums

Message 🔒

Statics

REQUEST_ID_COUNTER 🔒

Functions

handle_next_client_request 🔒

handle_request 🔒

Process a request asynchronously

handle_stdin_next_line 🔒

new_client 🔒

Constructor a JSON-RPC client. Returns a tuple: the client you can use to send requests, and the adapter you must pass to run_with_client() to connect the client to the proper IO.

run_with_client 🔒

Start doing JSON-RPC over stdio. The future will only return once stdin is closed or another IO error happens.

run_with_io 🔒